pyproject.toml
[tool.poetry]
name = "project-template"
version = "0.1.0"
description = "Basic Python FastAPi project structure for development."
authors = ["Sebastian Wdowiarz"]
readme = "README.md"
packages = []

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.104.1"
uvicorn = "0.27.1"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pre-commit-hooks = "4.5.0"
pre-commit = "3.5.0"

[tool.black]
line-length = 110
target-version = ['py39']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
